+++ NEWS.md
@@ -4,6 +4,7 @@
     * released 2023-10-25
     * sp_build_rsp_x update; x converted to data.frame (trouble with tibbles!)
     * pls_plot update; changed mar settings 
+    * added first draft pls_test 
 
 * [0.2.4] 
     * released 2023-10-14